@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <PreserveCompilationContext>true</PreserveCompilationContext>
     <AssemblyName>Smidge.Web</AssemblyName>
     <OutputType>Exe</OutputType>
@@ -16,22 +16,11 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Smidge\Smidge.csproj" />
-    <ProjectReference Include="..\Smidge.JavaScriptServices\Smidge.JavaScriptServices.csproj" />
     <ProjectReference Include="..\Smidge.Nuglify\Smidge.Nuglify.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="dotless" Version="1.6.7" />
-    <PackageReference Include="System.IO.Compression" Version="4.3.0" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="2.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.0.0" />
+    <Folder Include="Smidge\Cache\" />
   </ItemGroup>
-
-  <Target Name="PrepublishScript" BeforeTargets="PrepareForPublish">
-    <Exec Command="npm install" />
-  </Target>
 
 </Project>
