@@ -28,22 +28,53 @@ namespace Smidge.Nuglify
             services.AddSingleton<ISourceMapDeclaration, SourceMapDeclaration>();
 
             services.AddSingleton<NuglifySettings>(provider => nuglifySettings ?? new NuglifySettings(new NuglifyCodeSettings(null), new CssSettings()));
-            
+
+            services.AddTransient<IConfigureOptions<SmidgeOptions>, NuglifySmidgeOptionsSetup>();
+
+            //conventions
+            services.AddSingleton<IFileProcessingConvention, NuglifyMinifiedFilePathConvention>();
+
             return services;
         }
 
-        public static void UseSmidgeNuglify(this IApplicationBuilder app)
+        public static void UseSmidgeNuglify(this IApplicationBuilder app, bool useEndpointRouting = true)
         {
-            //Create custom route
-            app.UseMvc(routes =>
+#if NETCORE3_0
+            //NOTE: It's no longer polite to just call UseMVC as it enables things that the developer may 
+            //not need and the dev must disable EndpointRouting - so we let the dev decide.
+            //with core 3.0 you have to explicitly disable EndpointRouting se we default to on here 
+            if (useEndpointRouting)
             {
-                var options = app.ApplicationServices.GetRequiredService<IOptions<SmidgeOptions>>();
+                //Create custom route
+                app.UseEndpoints(endpoints =>
+                {
+                    var options = app.ApplicationServices.GetRequiredService<IOptions<SmidgeOptions>>();
 
-                routes.MapRoute(
-                    "SmidgeNuglifySourceMap",
-                    options.Value.UrlOptions.BundleFilePath + "/nmap/{bundle}",
-                    new {controller = "NuglifySourceMap", action = "SourceMap" });
-            });
+                    endpoints.MapControllerRoute(
+                        "SmidgeNuglifySourceMap",
+                        options.Value.UrlOptions.BundleFilePath + "/nmap/{bundle}",
+                        new { controller = "NuglifySourceMap", action = "SourceMap" });
+                });
+
+            }
+            else
+            {
+#endif
+                //Create custom route
+                app.UseMvc(routes =>
+                {
+                    var options = app.ApplicationServices.GetRequiredService<IOptions<SmidgeOptions>>();
+
+                    routes.MapRoute(
+                        "SmidgeNuglifySourceMap",
+                        options.Value.UrlOptions.BundleFilePath + "/nmap/{bundle}",
+                        new { controller = "NuglifySourceMap", action = "SourceMap" });
+                });
+#if NETCORE3_0
+            }
+#endif
+
+            
 
         }
     }
