@@ -6,9 +6,9 @@
             <li class="active"><a href="#">Runtime: @(Assembly.GetEntryAssembly().GetCustomAttribute<TargetFrameworkAttribute>().FrameworkName)</a></li>
         </ul>
         <ul class="left">
-            <li><a href="/"><i class="fa fa-camera-retro fa-1x"></i> Home</a></li>
-            <li><a href="/Home/SubFolder"><i class="fa fa-camera-retro fa-1x"></i> Sub page test</a></li>
-            <li><a href="/Home/AllDebug"><i class="fa fa-camera-retro fa-1x"></i> All debug</a></li>
+            <li><a href="~/"><i class="fa fa-camera-retro fa-1x"></i> Home</a></li>
+            <li><a href="~/Home/SubFolder"><i class="fa fa-camera-retro fa-1x"></i> Sub page test</a></li>
+            <li><a href="~/Home/AllDebug"><i class="fa fa-camera-retro fa-1x"></i> All debug</a></li>
         </ul>
     </section>
 </nav>
