@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using System.IO;
 
 namespace Smidge
 {
@@ -21,7 +22,11 @@ namespace Smidge
         /// Constructor that will use a smidge.json configuration file in env.ApplicationBasePath
         /// </summary>
         /// <param name="env"></param>
+#if NETCORE3_0
+        public SmidgeConfig(IWebHostEnvironment env)
+#else
         public SmidgeConfig(IHostingEnvironment env)
+#endif
         {
             //  use smidge.json file if it exists for backwards compatibility.
             var cfg = new ConfigurationBuilder()
@@ -33,7 +38,7 @@ namespace Smidge
         
         public string Version => _config["version"] ?? "1";
 
-        public string DataFolder => (_config["dataFolder"] ?? "App_Data/Smidge").Replace("/", "\\");
+        public string DataFolder => (_config["dataFolder"] ?? "Smidge").Replace('/', Path.DirectorySeparatorChar);
         
     }
 }
