@@ -18,15 +18,9 @@ namespace Smidge.Tests
         {
             var cssWithImport = @"@import url(""//fonts.googleapis.com/css?subset=latin,cyrillic-ext,latin-ext,cyrillic&family=Open+Sans+Condensed:300|Open+Sans:400,600,400italic,600italic|Merriweather:400,300,300italic,400italic,700,700italic|Roboto+Slab:400,300"");
 @import url(""//netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.css"");";
-                        
-            IEnumerable<string> importPaths;
-            var fileSystemHelper = new SmidgeFileSystem(Mock.Of<IFileProvider>(), Mock.Of<ICacheFileSystem>());
 
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-            var cssImportProcessor = new CssImportProcessor(fileSystemHelper, websiteInfo.Object, new RequestHelper(websiteInfo.Object));
-            var output = cssImportProcessor.ParseImportStatements(cssWithImport, out importPaths);
+            var cssImportProcessor = GetCssImportProcessor();
+            var output = cssImportProcessor.ParseImportStatements(cssWithImport, out _, out _);
 
             Assert.Equal(output, cssWithImport);
         }
@@ -35,24 +29,22 @@ namespace Smidge.Tests
         public void Can_Parse_Import_Statements()
         {
             var css = @"@import url('/css/typography.css');
-@import '/css/layout.css';
+@import '/css/layout.css' ;
+/*@import ""bootstrap/variables"";*/
 @import url('http://mysite/css/color.css');
-@import url(/css/blah.css);
+ @import url(/css/blah.css);
 @import ""css/blah2.css"";
 @import ""https://mysite.com/css/blah2.css"";
 
 body { color: black; }
 div {display: block;}";
 
-            IEnumerable<string> importPaths;
-            var fileSystemHelper = new SmidgeFileSystem(Mock.Of<IFileProvider>(), Mock.Of<ICacheFileSystem>());
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-            var cssImportProcessor = new CssImportProcessor(fileSystemHelper, websiteInfo.Object, new RequestHelper(websiteInfo.Object));
-            var output = cssImportProcessor.ParseImportStatements(css, out importPaths);
+            var cssImportProcessor = GetCssImportProcessor();
 
-            Assert.Equal(@"@import url('http://mysite/css/color.css');
+            var output = cssImportProcessor.ParseImportStatements(css, out IEnumerable<string> importPaths, out _);
+
+            Assert.Equal(@"/*@import ""bootstrap/variables"";*/
+@import url('http://mysite/css/color.css');
 body { color: black; }
 div {display: block;}".Replace("\r\n", string.Empty).Replace("\n", string.Empty), output.Replace("\r\n", string.Empty).Replace("\n", string.Empty));
 
@@ -62,6 +54,29 @@ div {display: block;}".Replace("\r\n", string.Empty).Replace("\n", string.Empty)
             //Assert.AreEqual("http://mysite/css/color.css", importPaths.ElementAt(2));
             Assert.Equal("/css/blah.css", importPaths.ElementAt(2));
             Assert.Equal("css/blah2.css", importPaths.ElementAt(3));
+        }
+
+        private CssImportProcessor GetCssImportProcessor()
+        {
+            var websiteInfo = GetWebsiteInfo();
+            var cssImportProcessor = new CssImportProcessor(GetFileSystem(), websiteInfo, new RequestHelper(websiteInfo));
+            return cssImportProcessor;
+        }
+
+        private ISmidgeFileSystem GetFileSystem()
+        {
+            var fileSystem = new SmidgeFileSystem(
+                Mock.Of<IFileProvider>(),
+                Mock.Of<ICacheFileSystem>());
+            return fileSystem;
+        }
+
+        private IWebsiteInfo GetWebsiteInfo()
+        {
+            var websiteInfo = new Mock<IWebsiteInfo>();
+            websiteInfo.Setup(x => x.GetBasePath()).Returns(string.Empty);
+            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
+            return websiteInfo.Object;
         }
     }
 }
