@@ -1,6 +1,5 @@
 ﻿using System;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 
 using Moq;
 using System.IO;
@@ -9,130 +8,20 @@ using Microsoft.Extensions.FileProviders;
 using Xunit;
 using Smidge.Hashing;
 using Smidge.Cache;
+using System.Threading.Tasks;
 
 namespace Smidge.Tests
 {
-
     public class FileSystemHelperTests
     {
-        [Fact]
-        public void No()
+        private ISmidgeFileSystem Create(string url = "~/Js/Test1.js")
         {
-
-            var url = "~/test/hello.js";
-
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-
-            var urlHelper = new RequestHelper(websiteInfo.Object);
-
-            var result = urlHelper.Content(url);
-
-            Assert.Equal("/test/hello.js", result);
-        }
-
-        [Fact]
-        public void Normalize_Web_Path_Virtual_Path()
-        {
-
-            var url = "~/test/hello.js";
-
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-
-            var urlHelper = new RequestHelper(websiteInfo.Object);
-           
-            var result = urlHelper.Content(url);
-
-            Assert.Equal("/test/hello.js", result);
-        }
-
-        [Fact]
-        public void Normalize_Web_Path_Relative()
-        {
-
-            var url = "test/hello.js";
-
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-
-            var urlHelper = new RequestHelper(websiteInfo.Object);
-
-            var result = urlHelper.Content(url);
-
-            Assert.Equal("test/hello.js", result);
-        }
-
-        [Fact]
-        public void Normalize_Web_Path_Absolute()
-        {
-
-            var url = "/test/hello.js";
-
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-
-            var urlHelper = new RequestHelper(websiteInfo.Object);
-
-            var result = urlHelper.Content(url);
-
-            Assert.Equal("/test/hello.js", result);
-        }
-
-        [Fact]
-        public void Normalize_Web_Path_External_Schemaless()
-        {
-
-            var url = "//test.com/hello.js";
-
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-
-            var urlHelper = new RequestHelper(websiteInfo.Object);
-
-            var result = urlHelper.Content(url);
-
-            Assert.Equal("http://test.com/hello.js", result);
-        }
-
-        [Fact]
-        public void Normalize_Web_Path_External_With_Schema()
-        {
-
-            var url = "http://test.com/hello.js";
-
-            var websiteInfo = new Mock<IWebsiteInfo>();
-            websiteInfo.Setup(x => x.GetBasePath()).Returns("/");
-            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
-
-            var urlHelper = new RequestHelper(websiteInfo.Object);
-
-
-            var result = urlHelper.Content(url);
-
-            Assert.Equal("http://test.com/hello.js", result);
-        }
-        
-        
-        
-
-        [Fact]
-        public void Get_File_Info_Non_Existent_File_Throws_Informative_Exception()
-        {
-
-            var webRootPath = "C:\\MySolution\\MyProject";
-
-            var url = "~/Js/Test1.js";
+            var webRootPath = $"C:{Path.DirectorySeparatorChar}MySolution{Path.DirectorySeparatorChar}MyProject";
 
             var cacheProvider = new Mock<ICacheFileSystem>();
             var fileProvider = new Mock<IFileProvider>();
             var file = new Mock<IFileInfo>();
-            string filePath = Path.Combine(webRootPath, "Js\\Test1.js");
+            string filePath = Path.Combine(webRootPath, $"Js{Path.DirectorySeparatorChar}Test1.js");
 
             file.Setup(x => x.Exists).Returns(false);
             file.Setup(x => x.IsDirectory).Returns(false);
@@ -147,6 +36,21 @@ namespace Smidge.Tests
                 fileProvider.Object,
                 cacheProvider.Object);
 
+
+            return helper;
+        }
+
+        [Fact]
+        public void Get_File_Info_Non_Existent_File_Throws_Informative_Exception()
+        {
+            var url = "~/Js/Test1.js";
+
+            var websiteInfo = new Mock<IWebsiteInfo>();
+            websiteInfo.Setup(x => x.GetBasePath()).Returns(string.Empty);
+            websiteInfo.Setup(x => x.GetBaseUrl()).Returns(new Uri("http://test.com"));
+
+            var helper = Create(url);
+
             FileNotFoundException ex = Assert.Throws<FileNotFoundException>(() => helper.SourceFileProvider.GetRequiredFileInfo(url));
 
             //    var result = helper.MapPath(url);
@@ -157,8 +61,8 @@ namespace Smidge.Tests
         [Fact]
         public void Reverse_Map_Path()
         {
-            var webRootPath = "C:\\MySolution\\MyProject";
-            var subPath = "Js\\Test1.js";
+            var webRootPath = $"C:{Path.DirectorySeparatorChar}MySolution{Path.DirectorySeparatorChar}MyProject";
+            var subPath = $"Js{Path.DirectorySeparatorChar}Test1.js";
             var filePath = Path.Combine(webRootPath, subPath);
 
             var file = new Mock<IFileInfo>();
@@ -177,6 +81,9 @@ namespace Smidge.Tests
                 cacheProvider.Object);
 
             var result = helper.ReverseMapPath(subPath, file.Object);
+
+            //Expected: ~/Js/Test1.js
+            //Actual:   ~/Js/Test1.js/Js\Test1.js
 
             Assert.Equal("~/Js/Test1.js", result);
         }
